Code migration from SpatialOS branch
- Added support for sending messages (chat)
- Added tracking of playerstates with events for UI to hook onto
- Removed AI Melee from C++ (moved to BP Task in upcoming submit)
- Fixed alignment of impact decal

Other commits taken from Spatial OS branch:
Simplified reading of the if-statement and removed authortity only flag (not needed and conflicts with spatial OS authority check this early in the startup loop)

Improved performance of time of day by replacing capture sky with curves.
Exposed SpawnNewBot as console command
Test timescale in GameStateBP to 0.0 to disable time of day for playtesting + disabled code that 'passifies' the bots during the day.
Minor changes for bot spawning